Rename file and remove unneeded file
