--- app/core/tests/security/seed_constants.py
@@ -1,16 +1,0 @@
-
-website_project = "website"
-people_depot_project = "people-depot"
-wanda_name = "Wanda"
-wally_name = "Wally"
-winona_name = "Winona"
-zani_name = "Zani"
-patti_name = "Patti"
-patrick_name = "Patrick"
-paul_name = "Paul"
-garry_name = "Garry"
-valerie_name = "Valerie"
-
-
-
-
